fix: 리뷰 반영 - #312
