+++ Projects/Feature/Home/Sources/Home/HomeContentSection.swift
@@ -16,19 +16,12 @@ import SharedPerfTestingSupport
 struct HomeContentSection: View {
     let store: StoreOf<HomeReducer>
 
-    #if PERF_TESTING
-    /// Pass 4-S2: 초기 layout settling 중 body가 여러 번 invalidation되어도 self-run scroll Task가
-    /// scene appearance당 한 번만 실행되도록 보호합니다.
-    @State private var selfRunScrollStarted: Bool = false
-    /// Pass 4-S2: scrollTo sequence가 끝나면 `"true"`로 바뀌며, trace 분석에서 post-scroll window를
-    /// 분리할 수 있도록 `perfStateMarker`로 노출합니다.
-    @State private var selfRunScrollDone: String = "false"
-    #endif
-
     var body: some View {
         #if PERF_TESTING
         if UITestMode.isEnabled, UITestMode.isSwiftUISelfRunFeedScroll {
-            selfRunScrollContent
+            HomeSelfRunFeedScrollHarness(store: store) {
+                scrollContent
+            }
         } else {
             scrollContent
         }
@@ -51,43 +44,6 @@ struct HomeContentSection: View {
             store.send(.fetchGoals)
         }
     }
-
-    #if PERF_TESTING
-    /// Pass 4-S2: Example/perf 전용 branch입니다.
-    /// 동일한 `scrollContent`를 `ScrollViewReader`로 감싸 launch-mode SwiftUI Template에서
-    /// interactive scroll attribution을 캡처할 수 있게 합니다.
-    private var selfRunScrollContent: some View {
-        ScrollViewReader { proxy in
-            scrollContent
-                .perfStateMarker(
-                    slug: "home",
-                    key: "swiftui-selfrun-scroll",
-                    value: selfRunScrollDone
-                )
-                .onAppear { startSelfRunScrollIfNeeded(proxy: proxy) }
-        }
-    }
-
-    private func startSelfRunScrollIfNeeded(proxy: ScrollViewProxy) {
-        guard !selfRunScrollStarted else { return }
-        selfRunScrollStarted = true
-        let allIds = store.items.map(\.id)
-        let stridedTargets = stride(from: 5, to: allIds.count, by: 5)
-            .compactMap { allIds.indices.contains($0) ? allIds[$0] : nil }
-        let preRollNanos: UInt64 = 1_000_000_000
-        let stepIntervalNanos: UInt64 = 300_000_000
-        Task { @MainActor in
-            try? await Task.sleep(nanoseconds: preRollNanos)
-            for id in stridedTargets {
-                withAnimation(.easeInOut(duration: 0.25)) {
-                    proxy.scrollTo(id, anchor: .top)
-                }
-                try? await Task.sleep(nanoseconds: stepIntervalNanos)
-            }
-            selfRunScrollDone = "true"
-        }
-    }
-    #endif
 
     var cardList: some View {
         LazyVStack(spacing: 16) {
