+++ Projects/Feature/Stats/Sources/Stats/StatsView.swift
@@ -15,17 +15,6 @@ import SharedPerfTestingSupport
 struct StatsView: View {
     @Bindable public var store: StoreOf<StatsReducer>
 
-    #if PERF_TESTING
-    /// Pass 4-S3 — guards the self-run scroll Task so it only fires once
-    /// per scene appearance, even if SwiftUI invalidates the view during
-    /// initial layout settling.
-    @State private var selfRunStatsScrollStarted: Bool = false
-    /// Pass 4-S3 — flips to `"true"` once the self-run scrollTo sequence
-    /// completes. Surfaced via `perfStateMarker` so trace analysis can
-    /// isolate the post-scroll window.
-    @State private var selfRunStatsScrollDone: String = "false"
-    #endif
-
     var body: some View {
         VStack(spacing: 0) {
             navigationBar
@@ -86,7 +75,9 @@ private extension StatsView {
     var cardList: some View {
         #if PERF_TESTING
         if UITestMode.isEnabled, UITestMode.isSwiftUISelfRunStatsScroll {
-            selfRunCardList
+            StatsSelfRunScrollHarness(items: store.items ?? []) {
+                scrollCardList
+            }
         } else {
             scrollCardList
         }
@@ -116,45 +107,6 @@ private extension StatsView {
         .background(Color.Gray.gray50)
     }
 
-    #if PERF_TESTING
-    /// Pass 4-S3 — Example/perf-only branch. Wraps the same
-    /// `scrollCardList` in a `ScrollViewReader` (public SwiftUI API) so a
-    /// self-running Task can call `proxy.scrollTo(id:anchor:)` across a
-    /// stride of `stats-heavy` item `goalId`s. State-driven self-run:
-    /// not equivalent to real finger drag; lower bound on real scroll cost.
-    private var selfRunCardList: some View {
-        ScrollViewReader { proxy in
-            scrollCardList
-                .perfStateMarker(
-                    slug: "stats",
-                    key: "swiftui-selfrun-scroll",
-                    value: selfRunStatsScrollDone
-                )
-                .onAppear { startSelfRunStatsScrollIfNeeded(proxy: proxy) }
-        }
-    }
-
-    private func startSelfRunStatsScrollIfNeeded(proxy: ScrollViewProxy) {
-        guard !selfRunStatsScrollStarted else { return }
-        selfRunStatsScrollStarted = true
-        let allIds = (store.items ?? []).map(\.goalId)
-        let stridedTargets = stride(from: 5, to: allIds.count, by: 5)
-            .compactMap { allIds.indices.contains($0) ? allIds[$0] : nil }
-        let preRollNanos: UInt64 = 1_000_000_000
-        let stepIntervalNanos: UInt64 = 300_000_000
-        Task { @MainActor in
-            try? await Task.sleep(nanoseconds: preRollNanos)
-            for id in stridedTargets {
-                withAnimation(.easeInOut(duration: 0.25)) {
-                    proxy.scrollTo(id, anchor: .top)
-                }
-                try? await Task.sleep(nanoseconds: stepIntervalNanos)
-            }
-            selfRunStatsScrollDone = "true"
-        }
-    }
-    #endif
-    
     var statsEmptyView: some View {
         Group {
             if store.isOngoing {
